+++ ebics-java-lib/src/main/java/org/ebics/client/xml/h004/InitializationRequestElement.java
@@ -86,23 +86,6 @@ public abstract class InitializationRequestElement extends DefaultEbicsRootEleme
   }
 
   /**
-   * Returns the digest value of the authenticated XML portions.
-   * @return  the digest value.
-   * @throws EbicsException Failed to retrieve the digest value.
-   */
-  public byte[] getDigest() throws EbicsException {
-    addNamespaceDecl("ds", "http://www.w3.org/2000/09/xmldsig#");
-
-    try {
-      return MessageDigest.getInstance("SHA-256", "BC").digest(Utils.canonize(toByteArray()));
-    } catch (NoSuchAlgorithmException e) {
-      throw new EbicsException(e.getMessage());
-    } catch (NoSuchProviderException e) {
-      throw new EbicsException(e.getMessage());
-    }
-  }
-
-  /**
    * Returns the element type.
    * @return the element type.
    */
