+++ ebics-java-lib/src/main/java/org/ebics/client/xml/h004/SPRRequestElement.java
@@ -22,6 +22,7 @@ package org.ebics.client.xml.h004;
 import org.ebics.client.api.EbicsSession;
 import org.ebics.client.exception.EbicsException;
 import org.ebics.client.order.EbicsAdminOrderType;
+import org.ebics.client.utils.CryptoUtils;
 import org.ebics.client.utils.Utils;
 import org.ebics.schema.h004.DataEncryptionInfoType.EncryptionPubKeyDigest;
 import org.ebics.schema.h004.*;
@@ -114,7 +115,7 @@ public class SPRRequestElement extends InitializationRequestElement {
     encryptionPubKeyDigest = EbicsXmlFactory.createEncryptionPubKeyDigest(session.getConfiguration().getEncryptionVersion(),
 								          "http://www.w3.org/2001/04/xmlenc#sha256",
 								          decodeHex(session.getBankCert().getE002Digest()));
-    signatureData = EbicsXmlFactory.createSignatureData(true, Utils.encrypt(Utils.zip(userSignature.prettyPrint()), keySpec));
+    signatureData = EbicsXmlFactory.createSignatureData(true, CryptoUtils.encrypt(Utils.zip(userSignature.prettyPrint()), keySpec));
     dataEncryptionInfo = EbicsXmlFactory.createDataEncryptionInfo(true,
 	                                                          encryptionPubKeyDigest,
 	                                                          generateTransactionKey());
