+++ ebics-java-lib/src/main/java/org/ebics/client/xml/h004/UploadInitializationRequestElement.java
@@ -25,6 +25,7 @@ import org.ebics.client.interfaces.ContentFactory;
 import org.ebics.client.io.Splitter;
 import org.ebics.client.order.AttributeType;
 import org.ebics.client.order.EbicsAdminOrderType;
+import org.ebics.client.utils.CryptoUtils;
 import org.ebics.client.utils.Utils;
 import org.ebics.schema.h004.DataEncryptionInfoType.EncryptionPubKeyDigest;
 import org.ebics.schema.h004.*;
@@ -165,7 +166,7 @@ public class UploadInitializationRequestElement extends InitializationRequestEle
     encryptionPubKeyDigest = EbicsXmlFactory.createEncryptionPubKeyDigest(session.getConfiguration().getEncryptionVersion(),
 								          "http://www.w3.org/2001/04/xmlenc#sha256",
 								          decodeHex(session.getBankCert().getE002Digest()));
-    signatureData = EbicsXmlFactory.createSignatureData(true, Utils.encrypt(Utils.zip(userSignature.prettyPrint()), keySpec));
+    signatureData = EbicsXmlFactory.createSignatureData(true, CryptoUtils.encrypt(Utils.zip(userSignature.prettyPrint()), keySpec));
     dataEncryptionInfo = EbicsXmlFactory.createDataEncryptionInfo(true,
 	                                                          encryptionPubKeyDigest,
 	                                                          generateTransactionKey());
