+++ ebics-java-lib/src/main/java/org/ebics/client/xml/h005/TransferRequestElement.java
@@ -80,23 +80,6 @@ public abstract class TransferRequestElement extends DefaultEbicsRootElement {
   }
 
   /**
-   * Returns the digest value of the authenticated XML portions.
-   * @return  the digest value.
-   * @throws EbicsException Failed to retrieve the digest value.
-   */
-  public byte[] getDigest() throws EbicsException {
-    addNamespaceDecl("ds", "http://www.w3.org/2000/09/xmldsig#");
-
-    try {
-      return MessageDigest.getInstance("SHA-256", "BC").digest(Utils.canonize(toByteArray()));
-    } catch (NoSuchAlgorithmException e) {
-      throw new EbicsException(e.getMessage());
-    } catch (NoSuchProviderException e) {
-      throw new EbicsException(e.getMessage());
-    }
-  }
-
-  /**
    * Returns the order type of the element.
    * @return the order type element.
    */
