+++ ebics-java-lib/src/main/java/org/ebics/client/xml/h005/UploadInitializationRequestElement.java
@@ -25,6 +25,7 @@ import org.ebics.client.interfaces.ContentFactory;
 import org.ebics.client.io.Splitter;
 import org.ebics.client.order.EbicsAdminOrderType;
 import org.ebics.client.order.h005.EbicsUploadOrder;
+import org.ebics.client.utils.CryptoUtils;
 import org.ebics.client.utils.Utils;
 import org.ebics.schema.h005.*;
 import org.ebics.schema.h005.DataEncryptionInfoType.EncryptionPubKeyDigest;
@@ -153,7 +154,7 @@ public class UploadInitializationRequestElement extends InitializationRequestEle
     encryptionPubKeyDigest = EbicsXmlFactory.createEncryptionPubKeyDigest(session.getConfiguration().getEncryptionVersion(),
 								          "http://www.w3.org/2001/04/xmlenc#sha256",
 								          decodeHex(session.getBankCert().getE002Digest()));
-    signatureData = EbicsXmlFactory.createSignatureData(true, Utils.encrypt(Utils.zip(userSignature.prettyPrint()), keySpec));
+    signatureData = EbicsXmlFactory.createSignatureData(true, CryptoUtils.encrypt(Utils.zip(userSignature.prettyPrint()), keySpec));
     dataEncryptionInfo = EbicsXmlFactory.createDataEncryptionInfo(true,
 	                                                          encryptionPubKeyDigest,
 	                                                          generateTransactionKey());
